refactor(design-system): move LiveCodeBlock to components folder
